+++ js/result-view.js
@@ -5,9 +5,6 @@ import {showResults} from '../js/show-results.js';
 import {MINIMUM_PLAYERS_LIVES, MINIMUM_PLAYER_TIME} from '../js/game.js';
 import {ResultMarkup} from '../js/result-markup.js';
 import {currentGame} from '../js/game-store.js';
-import {welcome} from '../js/welcome.js';
-import {showPage} from '../js/show-page.js';
-const STARTING_INDEX = 1;
 
 const className = `main main--result`;
 
@@ -58,8 +55,5 @@ export class ResultView extends AbstractView {
   }
 
   onMainReplayClickTry() {
-    currentGame.state.round = STARTING_INDEX;
-    currentGame.setInitialParams();
-    showPage(welcome);
   }
 }
