+++ js/welcome.js
@@ -1,6 +1,14 @@
 import {WelcomeView} from '../js/welcome-view.js';
+import {showRandomPage} from '../js/show-random-page.js';
 
 export const welcome = () => {
-  return new WelcomeView();
+  const welcomePage = new WelcomeView();
+
+  welcomePage.onMainPlayClick = () => {
+    showRandomPage();
+  };
+
+  return welcomePage;
 };
+
